+++ src/Execution/ReportingErrorHandler.php
@@ -33,10 +33,11 @@ class ReportingErrorHandler implements ErrorHandler
             return $next($error);
         }
 
-        $this->exceptionHandler->report(
-            // @phpstan-ignore-next-line TODO remove when supporting Laravel 7 and upwards
-            $error->getPrevious()
-        );
+        $previous = $error->getPrevious();
+        if ($previous !== null) {
+            // @phpstan-ignore-next-line Laravel versions prior to 7 are limited to accepting \Exception
+            $this->exceptionHandler->report($previous);
+        }
 
         return $next($error);
     }