+++ src/Schema/Directives/BuilderDirective.php
@@ -3,14 +3,15 @@
 namespace Nuwave\Lighthouse\Schema\Directives;
 
 use Nuwave\Lighthouse\Support\Contracts\ArgBuilderDirective;
+use Nuwave\Lighthouse\Support\Contracts\FieldBuilderDirective;
 
-class BuilderDirective extends BaseDirective implements ArgBuilderDirective
+class BuilderDirective extends BaseDirective implements ArgBuilderDirective, FieldBuilderDirective
 {
     public static function definition(): string
     {
         return /** @lang GraphQL */ <<<'GRAPHQL'
 """
-Use an argument to modify the query builder for a field.
+Manipulate the query builder with a method.
 """
 directive @builder(
   """
@@ -19,20 +20,34 @@ directive @builder(
   If you pass only a class name, the method name defaults to `__invoke`.
   """
   method: String!
-) repeatable on ARGUMENT_DEFINITION | INPUT_FIELD_DEFINITION
+
+  """
+  Pass a value to the method as the second argument after the query builder.
+  Only used when the directive is added on a field.
+  """
+  value: Mixed
+) repeatable on ARGUMENT_DEFINITION | INPUT_FIELD_DEFINITION | FIELD_DEFINITION
 GRAPHQL;
     }
 
-    /**
-     * Dynamically call a user-defined method to enhance the builder.
-     *
-     * @param  \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder  $builder
-     * @return \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder
-     */
     public function handleBuilder($builder, $value): object
     {
         $resolver = $this->getResolverFromArgument('method');
 
         return $resolver($builder, $value, $this->definitionNode);
+    }
+
+    public function handleFieldBuilder(object $builder): object
+    {
+        $resolver = $this->getResolverFromArgument('method');
+
+        if ($this->directiveHasArgument('value')) {
+            return $resolver(
+                $builder,
+                $this->directiveArgValue('value')
+            );
+        }
+
+        return $resolver($builder);
     }
 }