+++ src/Schema/TypeRegistry.php
@@ -253,7 +253,6 @@ EOL
         /** @var array<string, array<string, mixed>> $values */
         $values = [];
 
-        // @phpstan-ignore-next-line graphql-php types are unnecessarily nullable
         foreach ($enumDefinition->values as $enumValue) {
             /** @var \Nuwave\Lighthouse\Schema\Directives\EnumDirective|null $enumDirective */
             $enumDirective = $this->directiveFactory->exclusiveOfType($enumValue, EnumDirective::class);
@@ -346,8 +345,6 @@ EOL
                 $typeValue = new TypeValue($typeDefinition);
                 $fields = [];
 
-                // Might be a NodeList, so we can not use array_map()
-                // @phpstan-ignore-next-line graphql-php types are unnecessarily nullable
                 foreach ($typeDefinition->fields as $fieldDefinition) {
                     /** @var \Nuwave\Lighthouse\Schema\Factories\FieldFactory $fieldFactory */
                     $fieldFactory = app(FieldFactory::class);
@@ -371,7 +368,6 @@ EOL
                  * @return array<string, array<string, mixed>>
                  */
                 function () use ($inputDefinition): array {
-                    // @phpstan-ignore-next-line graphql-php types are unnecessarily nullable
                     return $this->argumentFactory->toTypeMap($inputDefinition->fields);
                 },
         ]);
@@ -387,7 +383,7 @@ EOL
             $typeResolver = $interfaceDirective->getResolverFromArgument('resolveType');
         } else {
             $typeResolver =
-                $this->findTypeResolverClass(
+                $this->typeResolverFromClass(
                     $nodeName,
                     (array) config('lighthouse.namespaces.interfaces')
                 )
@@ -405,7 +401,7 @@ EOL
     /**
      * @param  array<string>  $namespaces
      */
-    protected function findTypeResolverClass(string $nodeName, array $namespaces): ?Closure
+    protected function typeResolverFromClass(string $nodeName, array $namespaces): ?Closure
     {
         $className = Utils::namespaceClassname(
             $nodeName,
@@ -414,8 +410,10 @@ EOL
                 return method_exists($className, '__invoke');
             }
         );
+
         if ($className) {
             return Closure::fromCallable(
+                // @phpstan-ignore-next-line this works
                 [app($className), '__invoke']
             );
         }
@@ -447,7 +445,7 @@ EOL
             $typeResolver = $unionDirective->getResolverFromArgument('resolveType');
         } else {
             $typeResolver =
-                $this->findTypeResolverClass(
+                $this->typeResolverFromClass(
                     $nodeName,
                     (array) config('lighthouse.namespaces.unions')
                 )
@@ -464,8 +462,6 @@ EOL
                 function () use ($unionDefinition): array {
                     $types = [];
 
-                    // Might be a NodeList, so we can not use array_map()
-                    // @phpstan-ignore-next-line graphql-php types are unnecessarily nullable
                     foreach ($unionDefinition->types as $type) {
                         $types[] = $this->get($type->name->value);
                     }