+++ src/Subscriptions/Subscriber.php
@@ -62,6 +62,13 @@ class Subscriber implements Serializable
     public $args;
 
     /**
+     * The variables passed to the subscription query.
+     *
+     * @var array<string, mixed>
+     */
+    public $variables;
+
+    /**
      * The context passed to the query.
      *
      * @var \Nuwave\Lighthouse\Support\Contracts\GraphQLContext
@@ -79,6 +86,7 @@ class Subscriber implements Serializable
         $this->fieldName = $resolveInfo->fieldName;
         $this->channel = self::uniqueChannelName();
         $this->args = $args;
+        $this->variables = $resolveInfo->variableValues;
         $this->context = $context;
 
         /**
@@ -107,11 +115,19 @@ class Subscriber implements Serializable
 
         $this->channel = $data['channel'];
         $this->topic = $data['topic'];
-        $this->query = AST::fromArray( // @phpstan-ignore-line We know this will be exactly a DocumentNode and nothing else
+
+        /**
+         * We know the type since it is set during construction and serialized.
+         *
+         * @var \GraphQL\Language\AST\DocumentNode $documentNode
+         */
+        $documentNode = AST::fromArray(
             unserialize($data['query'])
         );
+        $this->query = $documentNode;
         $this->fieldName = $data['field_name'];
         $this->args = $data['args'];
+        $this->variables = $data['variables'];
         $this->context = $this->contextSerializer()->unserialize(
             $data['context']
         );
@@ -130,6 +146,7 @@ class Subscriber implements Serializable
             ),
             'field_name' => $this->fieldName,
             'args' => $this->args,
+            'variables' => $this->variables,
             'context' => $this->contextSerializer()->serialize($this->context),
         ]);
     }