+++ tests/Integration/Schema/Directives/HasManyDirectiveTest.php
@@ -466,7 +466,7 @@ class HasManyDirectiveTest extends DBTestCase
         $this->assertNotNull($user);
         /** @var array<string, mixed> $user */
         $tasks = Arr::first(
-            $user['fields'], // @phpstan-ignore-line
+            $user['fields'],
             function (array $field): bool {
                 return $field['name'] === 'tasks';
             }
@@ -735,11 +735,11 @@ class HasManyDirectiveTest extends DBTestCase
 
     public function testThrowsErrorWithUnknownTypeArg(): void
     {
-        $this->expectExceptionMessageRegExp('/^Found invalid pagination type/');
+        $this->expectExceptionMessage('Found invalid pagination type: foo');
 
         $schema = $this->buildSchemaWithPlaceholderQuery(/** @lang GraphQL */ '
         type User {
-            tasks(first: Int! after: Int): [Task!]! @hasMany(type:"foo")
+            tasks(first: Int! after: Int): [Task!]! @hasMany(type: "foo")
         }
 
         type Task {