+++ tests/Integration/Schema/Directives/SearchDirectiveTest.php
@@ -2,23 +2,26 @@
 
 namespace Tests\Integration\Schema\Directives;
 
-use Illuminate\Support\Collection;
+use Illuminate\Database\Eloquent\Collection as EloquentCollection;
+use Laravel\Scout\Builder as ScoutBuilder;
 use Laravel\Scout\EngineManager;
 use Laravel\Scout\Engines\NullEngine;
 use Mockery;
 use Mockery\MockInterface;
+use Nuwave\Lighthouse\Exceptions\DefinitionException;
+use Nuwave\Lighthouse\Scout\ScoutException;
 use Tests\DBTestCase;
 use Tests\Utils\Models\Post;
 
 class SearchDirectiveTest extends DBTestCase
 {
     /**
-     * @var \Mockery\MockInterface
+     * @var \Mockery\MockInterface&\Laravel\Scout\EngineManager
      */
     protected $engineManager;
 
     /**
-     * @var \Mockery\MockInterface
+     * @var \Mockery\MockInterface&\Laravel\Scout\Engines\NullEngine
      */
     protected $engine;
 
@@ -26,7 +29,7 @@ class SearchDirectiveTest extends DBTestCase
     {
         parent::setUp();
 
-        $this->engineManager = Mockery::mock();
+        $this->engineManager = Mockery::mock(EngineManager::class);
         $this->engine = Mockery
             ::mock(NullEngine::class)
             ->makePartial();
@@ -40,22 +43,24 @@ class SearchDirectiveTest extends DBTestCase
             ->andReturn($this->engine);
     }
 
-    public function testCanSearch(): void
+    public function testSearch(): void
     {
+        /** @var \Tests\Utils\Models\Post $postA */
         $postA = factory(Post::class)->create([
             'title' => 'great title',
         ]);
+        /** @var \Tests\Utils\Models\Post $postB */
         $postB = factory(Post::class)->create([
-            'title' => 'Really bad title',
+            'title' => 'Really great title',
         ]);
-        $postC = factory(Post::class)->create([
-            'title' => 'another great title',
+        factory(Post::class)->create([
+            'title' => 'bad title',
         ]);
 
         $this->engine
             ->shouldReceive('map')
             ->andReturn(
-                new Collection([$postA, $postC])
+                new EloquentCollection([$postA, $postB])
             );
 
         $this->schema = /** @lang GraphQL */ '
@@ -67,121 +72,284 @@ class SearchDirectiveTest extends DBTestCase
         type Query {
             posts(
                 search: String @search
-            ): [Post!]! @paginate
+            ): [Post!]! @all
         }
         ';
 
         $this->graphQL(/** @lang GraphQL */ '
         {
-            posts(first: 10 search: "great") {
-                data {
-                    id
-                    title
-                }
+            posts(search: "great") {
+                id
+                title
             }
         }
         ')->assertJson([
             'data' => [
                 'posts' => [
-                    'data' => [
-                        [
-                            'id' => $postA->id,
-                        ],
-                        [
-                            'id' => $postC->id,
-                        ],
+                    [
+                        'id' => $postA->id,
+                    ],
+                    [
+                        'id' => $postB->id,
                     ],
                 ],
             ],
         ]);
     }
 
-    public function testCanSearchWithCustomIndex(): void
+    public function testSearchWithEq(): void
     {
-        $postA = factory(Post::class)->create([
-            'title' => 'great title',
-        ]);
-        $postB = factory(Post::class)->create([
-            'title' => 'Really great title',
-        ]);
-        $postC = factory(Post::class)->create([
-            'title' => 'bad title',
-        ]);
+        $id = 1;
 
         $this->engine
             ->shouldReceive('map')
-            ->andReturn(
-                new Collection([$postA, $postB])
-            )
-            ->once();
-
-        $this->engine
-            ->shouldReceive('paginate')
-            ->with(
-                Mockery::on(
-                    function ($argument): bool {
-                        return $argument->index === 'my.index';
-                    }
-                ),
-                Mockery::any(),
-                Mockery::any()
-            )
-            ->andReturn(new Collection([$postA, $postB]))
+            ->withArgs(function (ScoutBuilder $builder) use ($id): bool {
+                return $builder->wheres === ['id' => $id];
+            })
+            ->andReturn(new EloquentCollection())
             ->once();
 
         $this->schema = /** @lang GraphQL */ '
         type Post {
-            id: ID!
-            title: String!
+            id: Int!
         }
 
         type Query {
             posts(
-                search: String @search(within: "my.index")
-            ): [Post!]! @paginate
+                id: Int @eq
+                search: String @search
+            ): [Post!]! @all
+        }
+        ';
+
+        $this->graphQL(/** @lang GraphQL */ '
+        query ($id: Int) {
+            posts(id: $id, search: "great") {
+                id
+            }
+        }
+        ', [
+            'id' => $id,
+        ])->assertJson([
+            'data' => [
+                'posts' => [],
+            ],
+        ]);
+    }
+
+    public function testSearchWithTrashed(): void
+    {
+        $this->engine
+            ->shouldReceive('map')
+            ->withArgs(function (ScoutBuilder $builder): bool {
+                return $builder->wheres === ['__soft_deleted' => 1];
+            })
+            ->andReturn(new EloquentCollection())
+            ->once();
+
+        $this->schema = /** @lang GraphQL */ '
+        type Post {
+            id: Int!
+        }
+
+        type Query {
+            posts(
+                id: Int @eq
+                search: String @search
+            ): [Post!]! @all @softDeletes
         }
         ';
 
         $this->graphQL(/** @lang GraphQL */ '
         {
-            posts(first: 10 search: "great") {
-                data {
-                    id
-                    title
-                }
+            posts(search: "foo", trashed: ONLY) {
+                id
+            }
+        }
+        ')->assertJson([
+            'data' => [
+                'posts' => [],
+            ],
+        ]);
+    }
+
+    public function testCanSearchWithinCustomIndex(): void
+    {
+        /** @var \Tests\Utils\Models\Post $postA */
+        $postA = factory(Post::class)->create([
+            'title' => 'great title',
+        ]);
+        /** @var \Tests\Utils\Models\Post $postB */
+        $postB = factory(Post::class)->create([
+            'title' => 'Really great title',
+        ]);
+        factory(Post::class)->create([
+            'title' => 'bad title',
+        ]);
+
+        $myIndex = 'my.index';
+
+        $this->engine
+            ->shouldReceive('map')
+            ->withArgs(function (ScoutBuilder $builder) use ($myIndex): bool {
+                return $builder->index === $myIndex;
+            })
+            ->andReturn(
+                new EloquentCollection([$postA, $postB])
+            )
+            ->once();
+
+        $this->schema = /** @lang GraphQL */ "
+        type Post {
+            id: ID!
+        }
+
+        type Query {
+            posts(
+                search: String @search(within: \"{$myIndex}\")
+            ): [Post!]! @all
+        }
+        ";
+
+        $this->graphQL(/** @lang GraphQL */ '
+        {
+            posts(search: "great") {
+                id
             }
         }
         ')->assertJson([
             'data' => [
                 'posts' => [
-                    'data' => [
-                        [
-                            'id' => "$postA->id",
-                        ],
-                        [
-                            'id' => "$postB->id",
-                        ],
+                    [
+                        'id' => $postA->id,
+                    ],
+                    [
+                        'id' => $postB->id,
                     ],
                 ],
             ],
         ]);
     }
 
+    public function testWithinMustBeString(): void
+    {
+        $this->schema = /** @lang GraphQL */ '
+        type Post {
+            id: ID!
+        }
+
+        type Query {
+            posts(
+                search: String @search(within: 123)
+            ): [Post!]! @all
+        }
+        ';
+
+        $this->expectException(DefinitionException::class);
+
+        $this->graphQL(/** @lang GraphQL */ '
+        {
+            posts(search: "great") {
+                id
+            }
+        }
+        ');
+    }
+
+    public function testMultipleSearchesAreNotAllowed(): void
+    {
+        $this->schema = /** @lang GraphQL */ '
+        type Post {
+            id: ID!
+        }
+
+        type Query {
+            posts(
+                first: String @search
+                second: String @search
+            ): [Post!]! @all
+        }
+        ';
+
+        $this->expectException(ScoutException::class);
+
+        $this->graphQL(/** @lang GraphQL */ '
+        {
+            posts(first: "great", second: "nope") {
+                id
+            }
+        }
+        ');
+    }
+
+    public function testIncompatibleArgBuildersAreNotAllowed(): void
+    {
+        $this->schema = /** @lang GraphQL */ '
+        type Post {
+            id: ID!
+        }
+
+        type Query {
+            posts(
+                search: String @search
+                nope: String @neq
+            ): [Post!]! @all
+        }
+        ';
+
+        $this->expectException(ScoutException::class);
+
+        $this->graphQL(/** @lang GraphQL */ '
+        {
+            posts(search: "great", nope: "nope") {
+                id
+            }
+        }
+        ');
+    }
+
+    public function testModelMustBeSearchable(): void
+    {
+        $this->schema = /** @lang GraphQL */ '
+        type Task {
+            id: ID!
+        }
+
+        type Query {
+            tasks(
+                search: String @search
+            ): [Task!]! @all
+        }
+        ';
+
+        $this->expectException(ScoutException::class);
+
+        $this->graphQL(/** @lang GraphQL */ '
+        {
+            tasks(search: "great") {
+                id
+            }
+        }
+        ');
+    }
+
     public function testHandlesScoutBuilderPaginationArguments(): void
     {
+        /** @var \Tests\Utils\Models\Post $postA */
         $postA = factory(Post::class)->create([
             'title' => 'great title',
         ]);
+        /** @var \Tests\Utils\Models\Post $postB */
         $postB = factory(Post::class)->create([
             'title' => 'Really great title',
         ]);
-        $postC = factory(Post::class)->create([
+        factory(Post::class)->create([
             'title' => 'bad title',
         ]);
 
         $this->engine->shouldReceive('map')
             ->andReturn(
-                new Collection([$postA, $postB])
+                new EloquentCollection([$postA, $postB])
             )
             ->once();
 
@@ -191,13 +359,12 @@ class SearchDirectiveTest extends DBTestCase
                 Mockery::any(),
                 Mockery::not('page')
             )
-            ->andReturn(new Collection([$postA, $postB]))
+            ->andReturn(new EloquentCollection([$postA, $postB]))
             ->once();
 
         $this->schema = /** @lang GraphQL */ '
         type Post {
             id: ID!
-            title: String!
         }
 
         type Query {
@@ -209,10 +376,9 @@ class SearchDirectiveTest extends DBTestCase
 
         $this->graphQL(/** @lang GraphQL */ '
         {
-            posts(first: 10 search: "great") {
+            posts(first: 10, search: "great") {
                 data {
                     id
-                    title
                 }
             }
         }