+++ tests/Unit/Schema/AST/DocumentASTTest.php
@@ -30,7 +30,7 @@ class DocumentASTTest extends TestCase
     public function testThrowsOnInvalidSchema(): void
     {
         $this->expectException(ParseException::class);
-        $this->expectExceptionMessageRegExp('/^Syntax Error/');
+        $this->expectExceptionMessage('Syntax Error: Unexpected Name "foo"');
 
         DocumentAST::fromSource('foo');
     }
@@ -81,7 +81,6 @@ class DocumentASTTest extends TestCase
 
         $this->assertInstanceOf(
             FieldDefinitionNode::class,
-            // @phpstan-ignore-next-line can not be null
             $queryType->fields[0]
         );