+++ tests/Unit/Schema/Directives/FieldDirectiveTest.php
@@ -106,7 +106,7 @@ class FieldDirectiveTest extends TestCase
     public function testThrowsAnErrorWhenNoClassFound(): void
     {
         $this->expectException(DefinitionException::class);
-        $this->expectExceptionMessage("No class 'NonExisting' was found for directive 'field'");
+        $this->expectExceptionMessage('No class `NonExisting` was found for directive `@field`');
 
         $this->schema = /** @lang GraphQL */ '
         type Query {